chore: morge plugin in app
Step 1 for removing the monorepo
