+++ packages/app/src/assets/ArrowDownwardIcon.js
