+++ packages/app/src/assets/ArrowUpwardIcon.js
