+++ packages/app/src/components/VisualizationPlugin/VisualizationPlugin.js
@@ -14,10 +14,10 @@ import { Button, IconLegend24, Layer } from '@dhis2/ui'
 import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React, { useEffect, useState, useCallback } from 'react'
-import { apiFetchLegendSets } from './api/legendSets.js'
+import { apiFetchLegendSets } from '../../api/legendSets.js'
+import { fetchData } from '../../modules/fetchData.js'
 import ChartPlugin from './ChartPlugin.js'
 import ContextualMenu from './ContextualMenu.js'
-import { fetchData } from './modules/fetchData.js'
 import PivotPlugin from './PivotPlugin.js'
 import styles from './styles/VisualizationPlugin.module.css'
