+++ packages/app/src/components/VisualizationPlugin/styles/PivotPlugin.style.js
