+++ packages/app/src/components/VisualizationPlugin/styles/VisualizationPlugin.module.css
