--- packages/plugin/src/index.js
@@ -1,1 +1,0 @@
-export { VisualizationPlugin as default } from './VisualizationPlugin.js'
